updates for universal
