+++ Frameworks/Plugins/ZappAppleVideoSubscriberSSO/Templates/ZappAppleVideoSubscriberSSO.podspec.ejs
@@ -17,10 +17,10 @@ Pod::Spec.new do |s|
   s.static_framework = true
   s.swift_versions = '5.1'
 
-  s.ios.frameworks = 'VideoSubscriberAccount'
+  s.frameworks = 'VideoSubscriberAccount'
 
-  s.ios.source_files  = [
-    'ios/**/*.swift'
+  s.source_files  = [
+    'Universal/**/*.swift'
   ]
 
   s.xcconfig =  {
